// perpx-frontend-clean/client/src/lib/eth/addresses.ts
// ⚠️ AUTO-GENERATED FILE
// DO NOT EDIT MANUALLY
// Generated from contracts/deployments/sepolia.json

export const CONTRACTS = {
  COLLATERAL_TOKEN: "0x8862eA92661e62947EdD15D6EcE4C375383032DD",
  PLP: "0xB82C2963B0D00A4DdFA94B8610f5eF44bBC53e45",
  LIQUIDITY_POOL: "0xF34A890015885C3b124F5c543b7516b3372eB3e1",
  PRICE_ORACLE: "0xcEb0e73B7c55077f23B162De392Cf8cD83fFB371",
  PERPETUAL_TRADING: "0x37a9052B8D2a30aBE419BBa36b550897E707940d",
  LIQUIDATION_ENGINE: "0xE37F6Fe19291322dccD6B0a3094228c64EC1A8a4",
  ROUTER: "0xf78bf671E4bcd6F4E08aBabA32a3629400631Ca7",
  CHAINLINK_ORACLE: "0xdfb85fdb2A7bdE95369434B62f8E5C7be2Ee61db",
} as const;